ETL/requirements.txt
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
cramjam==2.8.1
fastparquet==2024.2.0
fsspec==2024.2.0
idna==3.6
kaggle==1.6.6
logging==0.4.9.6
numpy==1.26.4
packaging==23.2
pandas==2.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
requests==2.31.0
six==1.16.0
text-unidecode==1.3
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
zipfile36==0.1.3
